+++ test/runtests.jl
@@ -277,9 +277,19 @@ for a in (w, x, y); test_approx_eq(erfcinv(erfc(a)), a); end
 for a in (w, x, y); test_approx_eq(erfcx(a), exp(a^2)*erfc(a)); end
 test_approx_eq(erfi(x), 1629.9946226015657 ± 914.3351093102547)
 for a in (w, x, y); test_approx_eq(dawson(a), 0.5*exp(-a^2)*erfi(a)*sqrt(pi)); end
+
+# Factorial and gamma
+test_approx_eq(factorial(x), Measurement(6, 0.7536706010590813))
+for a in (w, x, y); test_approx_eq(gamma(a), factorial(a - 1)); end
+for a in (w, x, y); test_approx_eq(gamma(a + 1), factorial(a)); end
+for a in (x, y); test_approx_eq(lgamma(a), log(gamma(a))); end
 test_approx_eq(digamma(y), 1.256117668431802 ± 0.056764591147422994)
 for a in (w, x, y); test_approx_eq(digamma(a), polygamma(0, a)); end
+for a in (w, x, y); test_approx_eq(digamma(invdigamma(a)), a); end
 test_approx_eq(polygamma(3, w), 193.40909103400242 ± 0.10422749480000776)
+for a in (w, x, y); test_approx_eq(trigamma(a), polygamma(1, a)); end
+
+# Beta function
 for a in (w, x, y); test_approx_eq(beta(a, x), gamma(a)*gamma(x)/gamma(a + x)); end
 for a in (w, x, y); test_approx_eq(beta(a, pi), gamma(a)*gamma(pi)/gamma(a + pi)); end
 for a in (w, x, y); test_approx_eq(beta(e, a), gamma(e)*gamma(a)/gamma(e + a)); end
@@ -287,11 +297,17 @@ for a in (x, y); test_approx_eq(lbeta(a, x), log(beta(a, x))); end
 for a in (x, y); test_approx_eq(lbeta(a, pi), log(beta(a, pi))); end
 for a in (x, y); test_approx_eq(lbeta(e, a), log(beta(e, a))); end
 
-# Factorial and gamma
-test_approx_eq(factorial(x), Measurement(6, 0.7536706010590813))
-for a in (w, x, y); test_approx_eq(gamma(a), factorial(a - 1)); end
-for a in (w, x, y); test_approx_eq(gamma(a + 1), factorial(a)); end
-for a in (x, y); test_approx_eq(lgamma(a), log(gamma(a))); end
+# Airy functions
+test_approx_eq(airy(0, x), 0.006591139357460721 ± 0.0011912976705951322)
+test_approx_eq(airy(1, x), -0.01191297670595132 ± 0.0019773418072382165)
+test_approx_eq(airy(2, x), 14.037328963730136 ± 2.2922214966382017)
+test_approx_eq(airy(3, x), 22.922214966382015 ± 4.211198689119041)
+for a in (w, x, y); test_approx_eq(airy(a), airy(0, a)); end
+for a in (w, x, y); test_approx_eq(airyai(a), airy(0, a)); end
+for a in (w, x, y); test_approx_eq(airyprime(a), airy(1, a)); end
+for a in (w, x, y); test_approx_eq(airyaiprime(a), airy(1, a)); end
+for a in (w, x, y); test_approx_eq(airybi(a), airy(2, a)); end
+for a in (w, x, y); test_approx_eq(airybiprime(a), airy(3, a)); end
 
 # Modulo
 let