+++ NEWS.md
@@ -5,7 +5,7 @@ New Features
 ------------
 
 * New mathematical operations supported: `erfi`, `dawson`, `digamma`,
-  `polygamma`.
+  `polygamma`, `beta`, `lbeta`.
 
 
 v0.1.0 (2016-06-18)